Fixed 2nd set of pull request issues, added dependencies to allow custom ForkLift msg to work
