+++ msg/ForkLift.msg
@@ -1,3 +1,6 @@
+# Height of the forklift above the ground, in millimeters
 float64 height
+# The ratio from 0 to 1 that the forklift is raised, 0 at the bottom, 1 at the top
 float64 ratio
+# The angle of the forklift arm relative to the ground, in radians
 float64 angle
